DefangIpAddress.cs optimizations.
Start for loop at address.Length - 2 since the last character can never be a '.' in a valid address.
Decrement i every time a '.' is found.  Combined with the decrement you get in the for loop this will make sure we skip checking the number and go straight to where the next '.' should be.

These changes depend on the address given being valid which the question says it will be.